feat: start work with a mocking setup for our cypress tests
