+++ .env.test
@@ -1,1 +1,2 @@
 AUTH0_BASE_URL="http://localhost:3000"
+MONGODB_URI="mongodb://127.0.0.1:27017/toiletmap"
