+++ package.json
@@ -9,7 +9,7 @@
     "start": "next start",
     "cypress": "cypress open",
     "cypress:headless": "cypress run",
-    "e2e": "NODE_ENV=test yarn build && start-server-and-test start http://localhost:3000 cypress",
+    "e2e": "NODE_ENV=test yarn build && NODE_ENV=test start-server-and-test start http://localhost:3000 cypress",
     "e2e:headless": "NODE_ENV=test yarn build && start-server-and-test start http://localhost:3000 cypress:headless",
     "lint": "next lint",
     "check": "tsc -p tsconfig.json --noEmit",
@@ -18,6 +18,7 @@
     "migrate": "migrate-mongo up",
     "generate": "node -r dotenv/config ./src/api/db/manage/generate/index.js --confirm",
     "areas": "node -r dotenv/config ./src/api/db/manage/areaToDatabase/index.js",
+    "generateMocks": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' ts-node --transpileOnly ./scripts/generateMocks.ts",
     "prepare": "husky install && cp ./node_modules/graphql-voyager/dist/voyager.worker.js ./public && yarn codegen",
     "codegen": "graphql-codegen --config codegen.yml",
     "start-storybook": "start-storybook -p 6006",
@@ -74,6 +75,7 @@
     "next": "^12.1.5",
     "next-plausible": "^3.1.9",
     "ngeohash": "^0.6.3",
+    "nodemon": "^2.0.15",
     "prop-types": "^15.8.1",
     "query-string": "^7.1.1",
     "react": "^18.0.0",
@@ -90,6 +92,7 @@
     "@commitlint/cli": "16.2.3",
     "@commitlint/config-conventional": "^16.2.1",
     "@emotion/babel-preset-css-prop": "^11.2.0",
+    "@faker-js/faker": "^6.2.0",
     "@graphql-codegen/cli": "^2.6.2",
     "@graphql-codegen/import-types-preset": "^2.1.15",
     "@graphql-codegen/typescript": "^2.4.8",
@@ -111,7 +114,7 @@
     "@types/leaflet.markercluster": "^1.4.6",
     "@types/lodash": "^4.14.182",
     "@types/ngeohash": "^0.6.4",
-    "@types/node": "^17.0.25",
+    "@types/node": "^17.0.26",
     "@types/react": "^18.0.5",
     "@types/react-dom": "^18.0.1",
     "@types/react-leaflet": "^2.8.2",
@@ -135,6 +138,7 @@
     "jest": "^27.5.1",
     "lint-staged": "^12.4.0",
     "migrate-mongo": "^9.0.0",
+    "mongodb-memory-server": "^8.5.1",
     "prettier": "^2.6.2",
     "start-server-and-test": "^1.14.0",
     "storybook-addon-apollo-client": "^4.0.11",
