+++ src/api/db/manage/areaToDatabase/config.json
@@ -3,8 +3,8 @@
     {
       "id": 1,
       "src": {
-        "remote": true,
-        "url": "https://opendata.arcgis.com/datasets/0e07a8196454415eab18c40a54dfbbef_0.geojson"
+        "remote": false,
+        "url": "authbound.geojson"
       },
       "version": 2,
       "priority": 5,
