+++ src/api/resolvers.ts
@@ -136,6 +136,7 @@ const resolvers = {
             ],
           ],
         });
+
       return stringifyAndCompressLoos(loos);
     },
     areas: async () => {
